Boards: Templates: fix warnings and errors
Fixed ARMCLANG and GCC compiler warnings and errors of Baremetal apps

+++ Boards/Templates/Baremetal/CANFD_Bus_Monitor.c
@@ -21,6 +21,7 @@
  ******************************************************************************/
 
 #include <stdio.h>
+#include <inttypes.h>
 #include <RTE_Components.h>
 #include CMSIS_device_header
 #include "pinconf.h"
@@ -215,7 +216,7 @@ static void canfd_lom_demo_thread_entry(void)
                                               &service_error_code);
     if(error_code)
     {
-        printf("SE Error: HFOSC clk enable = %d\n", (int)error_code);
+        printf("SE Error: HFOSC clk enable = %"PRIu32"\n", error_code);
         return;
     }
 
@@ -226,7 +227,7 @@ static void canfd_lom_demo_thread_entry(void)
                                               &service_error_code);
     if(error_code)
     {
-        printf("SE Error: 160 MHz clk enable = %d\n", (int)error_code);
+        printf("SE Error: 160 MHz clk enable = %"PRIu32"\n", error_code);
         return;
     }
 
@@ -238,7 +239,7 @@ static void canfd_lom_demo_thread_entry(void)
     ret_val = board_pins_config();
     if (ret_val != 0)
     {
-        printf("Error in pin-mux configuration: %d\n", ret_val);
+        printf("Error in pin-mux configuration: %"PRId32"\n", ret_val);
         return;
     }
 
@@ -250,14 +251,14 @@ static void canfd_lom_demo_thread_entry(void)
     ret_val = board_canfd_pins_config();
     if(ret_val != 0)
     {
-        printf("Error in pin-mux configuration: %d\n", ret_val);
+        printf("Error in pin-mux configuration: %"PRId32"\n", ret_val);
         return;
     }
 #endif
 
     /* Get CANFD capabilities */
     can_capabilities = CANFD_instance->GetCapabilities();
-    printf("Num of objects supported: %d\r\n", can_capabilities.num_objects);
+    printf("Num of objects supported: %"PRIu8"\r\n", can_capabilities.num_objects);
 
     /* Initializing CANFD Access struct */
     ret_val = CANFD_instance->Initialize(cb_unit_event, cb_object_event);
@@ -383,7 +384,7 @@ uninitialise_canfd:
                                               &service_error_code);
     if(error_code)
     {
-        printf("SE Error: HFOSC clk disable = %d\n", (int)error_code);
+        printf("SE Error: HFOSC clk disable = %"PRIu32"\n", error_code);
         return;
     }
 
@@ -394,7 +395,7 @@ uninitialise_canfd:
                                               &service_error_code);
     if(error_code)
     {
-        printf("SE Error: 160 MHz clk disable = %d\n", (int)error_code);
+        printf("SE Error: 160 MHz clk disable = %"PRIu32"\n", error_code);
         return;
     }
 
@@ -459,7 +460,7 @@ static void canfd_check_error(void)
         }
         else
         {
-            printf("Error in CANFD-->Error code: %d\r\n",
+            printf("Error in CANFD-->Error code: %"PRIu8"\r\n",
                     cur_sts.last_error_code);
         }
         bus_error = false;
@@ -505,8 +506,8 @@ static void canfd_process_rx_message(void)
          * performs the below operations */
         if(rx_msg_header.rtr == 1U)
         {
-            printf("Rx msg:\r\n    Type:Remote frame, Id:%lu",
-                   (rx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)));
+            printf("Rx msg:\r\n    Type:Remote frame, Id:%"PRIu32"",
+                   (uint32_t)(rx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)));
         }
         else
         {
@@ -521,8 +522,8 @@ static void canfd_process_rx_message(void)
                     printf("\r\n    Error Occurred in Rx message \r\n");
                     return;
                 }
-                printf("Id:%lu, Len:%d:\r\n    Data:",
-                       (rx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)),
+                printf("Id:%"PRIu32", Len:%"PRIu8":\r\n    Data:",
+                       (uint32_t)(rx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)),
                        rx_msg_size);
                 for(iter = 0; iter < rx_msg_size; iter++)
                 {
